requirements.txt
pytrec_eval
numba
h5py
datasets
pytorch_lightning==1.3.2
transformers==4.24.0
Pillow
tqdm==4.56.0
ipdb==0.13.4
numpy==1.19.5
einops==0.3.0
pyarrow==2.0.0
sacred==0.8.2
pandas
torchmetrics==0.6.0
ftfy